backend/requirements.txt
flask
flask-cors
python-dotenv
google-genai
openai
PyPDF2